[BSE-2302] Support type syntax TIMESTAMP WITH TIME ZONE (#6975)
Supporting the aforementioned syntax in the context `to_char(CURRENT_DATE::timestamp with time zone, 'YYYYMMDD'::text)`. We do not yet have a proper `TIMESTAMP_TZ` type, so this PR adds the parser support but aliases it to `TIMESTAMP_LTZ`, which is still correct for the purposes of the query it is being used in.