[BSE-2477] Partial fix ambigous colname issue (#7095)
This is a partial fix that resolves two of the ~20 USING bug queries. Essentially, we weren't calling a helper function that handles expanding USING clause in a specific case. It was a very easy fix. I'm doing this separately because it's easy to do so, and may fix other issues.