[BSE-2564] Coalesce Type resolution fix (#7146)
This PR fixes an incorrect insert in the RexSimplifier code, and adds some tests that will be unblocked by a subsequent PR.

Our projection push down code does not update the null ability of Rex calls when they are pushed past a left/right join. While this is a missed optimization, this is actually allowed behavior. Therefore the assert inside of the rexSimplifier code is incorrect.