[BSE-2588] Fix codegen error with join on equality between input ref and a string (#7143)
The query concentric_property amongst others has a join condition like this: PandasJoin(condition=[AND(=($1, $2), =('fcf2d328-e90e-44c9-ab9c-3431669a6cc3', $3))], joinType=[left])

The problem is our codegen for the non-equijoin condition of comparing a column from the RHS to a string literal. We have a section of code checking for merged columns when doing =, but it only checks if the 2nd argument is a RexInputRef before proceeding with the logic. We need to also check the first argument, since in cases like this it will throw an error when it thinks that the codegen emitted for the string literal corresponds to a column name.