Fixing test_option_array_contains Nightly Failure (#7201)
Fixing the test slow test_option_array_contains which broke once we merged a PR that enforced stronger requirements on the inputs to semi-structured array functions in line with how we expect them to be used by BodoSQL. The old version of this test passed in a list whereas we should have passed in an array.

Also fixing the failure from test_coerce_scalar_to_array_unknown which comes from not prohibiting the test on multiple ranks. We should be doing so on this test since it is designed to return a singleton Series whose value is checked back in Python.