[BSE-2626] Support TO_DOUBLE on variants (#7216)
Adds support for TO_DOUBLE to be called on variants. The type handling is done on the Bodo side, and can accept any integer/float/string/boolean columns. Notably, boolean support is added to the Bodo implement even though TO_DOUBLE cannot be called directly on booleans because it can be called on variants containing booleans (False becomes 0.0, True becomes 1.0).