[BSE-2516] Fix join condition with non correlated IN subquery as the condition (#7209)
The Calcite SubQueryRemovalProgram has a bug on joins conditions like this:
```sql
SELECT ...
FROM T1
INNER JOIN T2
ON T1.A IN (subquery)
```
Where the subquery is non-correlated. The bug has to do with the handling of the two inputs (`T1` and `T2`) and the handling of the number of columns being incorrect versus when the same `IN` expansion is done with a single RelNode as an input. The fix made by this PR is to eject the `IN` from the join condition to a parent filter when pulling the filter upwards is valid.