[BSE-2070] Fix logical correlates not removed from plan due to FLATTEN inside IN (#7228)
Fixes bug in decorrelation that prevents some logical correlates from being removed. The bug was caused when a flatten call was inside of an IN clause, since we previosuly did subquery removal after the conversion of flatten from table function scan to a logical flatten node. Switching the order fixes this, but the affected query is not testable because it takes an ungodly amount of time for the planner to work on it (due to re-used common table expressions making the plan explode exponentially in size).

Ancillary change: altered testing infrastructure to set the default charset before any query plan/codegen tests to prevent rel conversion errors for certain string literals.