[BSE-2637] Updates error message for nested udf resolving (#7269)
Fixes an issue with UDF resolution. This issue cannot be reproduced in the maven test suite. There's an open PR to determine why this issues exists when running the actual test vs the test suite, and to create a maven reproducer of this issue. There's a followup issue with my brief investigation/thoughts here: https://bodo.atlassian.net/browse/BSE-2770