[BSE-1507] PARSE_URL support (#7211)
This implementation is still extremely bugging, and doesn't match Snowflake's behavior for several edge cases. There are ~6 followup issues, they're commented in the code where relevant. I do want to get this merged, because even in this state, we may be able to unblock the specific target customer queries. If any of the reviewers feel otherwise, I'm happy to leave this as an open PR until I handle the followups.