[BSE-2638]  Fix for error of dry-stage query (#7277)
The issue essentially stems from the fact that we were leaving a SqlNode in an incorrect state as a result of our qualify handling. This led to issues when the associated SqlNode was used in several locations (because of a WITH statement).

After fixing this error, the query now stack overflows. I'm going to look into this as a followup: https://bodo.atlassian.net/browse/BSE-2821