[BSE-3019] Add caching for SQL plans in Bodo compilation process (#7508)
(As title)

Added a `BodoSqlPlanCache` class which helps with caching plans generated while compiling a Bodo function.
Modified typing pass to get the SQL plan string in addition to the generated Pandas code when compiling `bc.sql` statements. To enable this, we added a couple of simple APIs in `RelationalAlgebraGenerator`. We also added a simple `PandasCodeSqlPlanPair` class in Java to store a pair of strings.
We then use `BodoSqlPlanCache` to save it to disk. Note that at this point we only write the plans to disk and we never load them from disk for future use. The idea is that the client, e.g. BodoSQLWrapper on the Platform, can get the plan after the query is compiled. `BodoSqlPlanCache` exposes an API to get the expected location of where the plan would be cached based on the SQL query text. This makes the process deterministic.