[BSE-3109] Fixes issue with rel subsets when pushing down iceberg filters/limits (#7559)
Previously, we avoided pushing filters into iceberg if there was already a limit present in the iceberg tree. This caused issues, because during the volcano plan, all the nodes will be relSubsets, leading to a casting error. To resolve this, we now allow pushing multiple limits into the iceberg tree, and resolve them in flattening by taking the min. This change will result in plans that handle this edgecase inefficiently, but as this is a fairly rare edgecase, we think it's fine for now.