BSE-3246: Fix stack overflow when generating partial filter pushdown condition (#7678)
Previously generating the partial filter pushdown condition for iceberg filters could lead to a stack overflow due to a check that inserted a not null filter even after the not null filter was already inserted in a previous pass. Now we union the previous condition with the candidate condition and only use the candidate condition if they do not simplify to the same thing.