Fixing _bodo_read_as_dict error in Azure Nightly (#7682)
Fixing an error caused by adding _bodo_read_as_dict to all of our Iceberg Read BodoSQL code. The function did not work as intended when _bodo_read_as_dict was passed in as None even though None is the default, so now this makes a minor change to ensure it is safe to pass in _bodo_read_as_dict=None.