[BSE-3265] Change query profile output mechanism (#7689)
Currently codegen assumes that the default value for `BODO_TRACING_LEVEL` is 1, and 0 will disable tracing codegen. However, `QueryProfileCollector::Finalize` assumes a default of 0, and setting to 1 will enable outputting profiles. This is awkward and confusing to use, so this PR set the default tracing level to 1 everywhere, and requires setting`BODO_TRACING_OUTPUT_DIR` to generate output. Additionally, profiles are now written to `<BODO_TRACING_OUTPUT_DIR>/run_<TIMESTAMP>/query_profile*.json` to avoid the requirement of randomized directory names in general.