Skip build/probe shuffle buffer reserve when shuffle is not possible (#7714)
(As title)

While going through the Query Profiles, I saw that there was some time spent in "Appending" to the probe shuffle buffer in a broadcast join case. Of course, nothing was appended, but because we made the "AppendBatch" call, it was still reserving space, which took some time. This just skips that step entirely when shuffle isn't possible.