Add a specialized pre-aggregation hash-table key comparator when all key types are the same (#7763)
(As title)

In `get_group_info`, which is used during pre-aggregation in the streaming case and also in non-streaming, we use a templated equality comparator for the hash map in certain special cases such as 1 key, 2 keys, etc. This adds a new case where we use a templated comparator in the case where all key types are the same. When there are 4 or more keys of the same type, we also add a short-circuit step by comparing the hashes before the equality comparisons.
For the query, the 1st change improved the performance of the grouping step of the pre-aggregation stage of the expensive Aggregate by ~22%. The second change improved the performance by a further ~4%. In total, this change improved the performance of that step by ~25%.

The PR also makes some minor changes to the decimal multiplication code such as adding `noexcept` and `inline` keywords to many of the functions.