Add more complex join tests to stress-test nested join caching (#7811)
These tests are inspired by the edge cases that caused unexpected caching failures when other planner changes were made. The tests introduce new fake schemas with a colleciton of tables that can be used in a similar fashion to TPCH, but with more of a "star schema" layout instead of a "snowflake schema" layout. See here for the distinction: https://www.thoughtspot.com/data-trends/data-modeling/star-schema-vs-snowflake-schema