[BSE-3475] Do not allocate 2x frames if already setting memory as 5x (#7898)
(As title)

For context, the reason we ran into this issue is that when spilling is not available, we allocate 5x the amount of system memory. In this particular case, since there’s a single process and the node has 1TiB memory, that single process will try to allocate a lot of virtual address space. 
If we do `px python -c "import bodo; print(bodo.__version__)"`, it doesn’t error out since each process is allocating a lot fewer addresses.
In the single process case, it will try to allocate ~17x5 TiB of virtual address space (93,458,488,360,960 addresses). With my previous PR, that allocates extra frames, this goes up to 29 x 5TiB = 159,429,186,027,520 addresses. That is because we statically add (8+ (8/2)) SizeClass-es worth of extra space.
In 64-bit systems, only the top 48 bits are usually used and that gives us a theoretical maximum of 281,474,976,710,656. OS-es typically reserve half of that, leaving us with ~140T addresses. This explains why trying to allocate ~160T addresses throws an error.
Given this, the most robust solution might be to ensure we never try to allocate >2^47 addresses. In fact, to be safe, we should probably do 2^45. I’ll think more about how to do this and what the appropriate bounds are and open a follow-up PR.


Verified on the platform that this fixes the issue:
```sh
(DEV) bodo@ip-10-30-3-112: ~ $ BODO_BUFFER_POOL_DISABLE_SPILLING=1 python -c "import bodo; print(bodo.__version__)"
2024.7.dev46+gcded460825
```