[BSE-3513] Support low-ndv IN filters for Snowflake I/O runtime join filters (#7928)
Builds up a hash set of unique values on the build side of a join, until the size goes over a certain threshold that can be controlled by an environment variable (default=20). This hashset is used for numeric types to produce an I/O runtime join filter on the precise set of unique values, as opposed to just the min/max.