[BSE-3718]/[BSE-3683] Add support for decimal sum without groupby or with window in windowed_sum and decomposes var/std/covar/corr as window functions into sum/count (#8146)
Extending `SUM` support on decimals to include no-groupby aggregation, and also the windowed_sum kernel which is used to calculate `SUM` as a window function in non-streaming contexts. This is a pre-requisite for adding decimal support for moment functions via decomposing into sum, at least until the no-groupby and non-streaming window paths are changed.

Also decomposes the functions mentioned into combinations of sum/count, unless they are called with `DISTINCT`. In doing so, we gain the following benefits:
- Increased opportunities for re-use of computations when multiple window functions are called together.
- Automatic decimal coverage for all of these window functions.
- Creating a fast-path to streaming coverage for all of these functions, since we only require sum/count to be implemented in a particular scenario in order for all of these functions to work.
Main downside: creates the full answers and broadcasts them even for `OVER ()` on these specific functions, which is much less performant than doing the computations when everything has been compressed down to a single row of data. We can switch some of these functions to avoid doing rewrites once we add `OVER ()` support for them.