[BSE-3744] Decompose PERCENT_RANK, CUME_DIST and RATIO_TO_REPORT in the planner (#8231)
In doing so, implicitly enables streaming for RATIO_TO_REPORT. Also, this ensures we do not need to re-write PERCENT_RANK in the new sort-based implementation. Made a bunch of ancillary changes to make the defaults of frames be ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING when it makes sense, and also explicitly enforcing window frame bans from Snowflake on most functions.