Only create/replace catalog integration if it doesn't already exist (when coverting Iceberg table to SF-managed) (#8247)
(As title)

This is important in case there are failures between registering the table in the temporary catalog and converting it to a SF-managed table. If we fail between these steps, we end up with tables that are tied to this temporary catalog integration object. During the next write, the "CREATE OR REPLACE" will fail because it cannot drop (for replacement) a catalog integration with tables tied to it. Using `CREATE IF NOT EXISTS` only creates it if it doesn't already exist and doesn't try to replace it.
Verified that it works as expected using `test_snowflake_iceberg_catalog.py::test_snowflake_catalog_iceberg_write`.

Also modifies the call to `get_table_target_file_size_bytes` to call it with `is_parallel=True`. Without it, we get warning like:
```
/home/bodo/Bodo/bodo/io/stream_iceberg_write.py:846: UserWarning: Got exception while trying to get write.target-file-size-bytes: Rank 11 is trying to launch the JVM. Only rank 0 should launch it., continuing with default value
```
which can be confusing for users. Verified that it works by running `test_stream_iceberg_write.py::test_iceberg_write_basic_rep`.