[BSE-3892] Fix a bug in `concat_tables` that accidentally unpins the input even if it is originally pinned (#8314)
(As title)

During `concat_tables`, we call `ReserveTable` which assumes that the input is unpinned. For string arrays, we pin the array (assuming it’s unpinned), get the length, and then unpin it back. When the input is pinned, we end up with an unpinned array, which then fails during the `UnsafeAppendBatch` step since we’re reading garbage data. In particular, for strings, we call `n_sub_elems()` to get the length of the array, which will return garbage data.