[BSE-3972] Fix Micromamba in Github Actions E2E Tests (#8348)
Fix the Micromamba version used in our E2E tests to 1.5.0-0 because the latest version (2.0.0) has a problem with Github Actions. It looks like the micromamba team is working on fixing it, but we don't need the new version necessarily.

Note, the reason we didn't notice the problem before was because we were loading the conda environment from cache. On Friday, we updated the lock file so it manually constructed the environment.

Successful E2E Run: https://github.com/Bodo-inc/Bodo/actions/runs/11118465491/job/30892085457