[BSE-3990] Modify Aggregate `COUNT()` to not require input columns (#8364)
This PR allows the `size` aggregation to have no input column since they're technically not required and we were passing the first key column as a dummy. To do this, we create a new `SizeColSet`, which doesn't require an input column and computes the output based on the `grouping_info`. The BodoSQL codegen has been modified to match this and corresponding changes have been made in the compiler not to need any input columns. The non-streaming code path may still pass input columns, but they will simply be ignored at runtime.