[BSE-4138] Debug snowflake e2e tests (#8574)
The issue was that the user program creates this lazy "tmp" directory for storing the sas token, but it's lazy so it doesn't exist yet.

The workers then inherit this environment variable from the spawner which tells them where to look for the sas token. However, at the same time the workers also create their own directory for the sas token and put it there.

This issue is since they inherit the environment variable from the spawner, they end up looking for the sas token in a directory that does not exist.