Execute plan before copying Series (#497)
As titled. Copying a Series does not preserve the index because the manager copies the axes first, which doesn't trigger an execute/collect. See pandas/core/managers.py:557 for more details.

Note that this doesn't change whether we collect in copy or not, it just changes when the collect happens to make sure the attributes needed by copy are correct up front.